Splitted Person model logic into Person repository
